+++ mobile/android/android-components/docs/changelog.md
@@ -14,6 +14,9 @@ permalink: /changelog/
 * **lib/publicsuffixlist**
   * ⚠️ **This is a breaking change**: Removed `String.urlToTrimmedHost` extension method.
 
+* **feature-top-sites**
+  * Extend `DefaultTopSitesStorage` to accept a `TopSitesProvider` for fetching top sites. [#11483](https://github.com/mozilla-mobile/android-components/issues/11483)
+
 # 97.0.0
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v96.0.0...v97.0.0)
 * [Milestone](https://github.com/mozilla-mobile/android-components/milestone/144?closed=1)